requirements.txt
discord.py
google_api_python_client
google_auth_oauthlib
protobuf
python-dotenv
google-generativeai